services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
pyyaml==6.0
passlib[bcrypt]==1.7.4